fix(cloudflare): address zaraz review
